requirements.txt
torch>=2.2.0
transformers>=4.45.0
accelerate
qwen-vl-utils
datasets
Pillow>=10.0.0
